requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
easyocr==1.7.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
lxml==5.2.2
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
newspaper3k==0.2.8
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pyclipper==1.3.0.post5
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
safetensors==0.4.3
scikit-image==0.21.0
scipy==1.10.1
sgmllib3k==1.0.0
shapely==2.0.5
six==1.16.0
soupsieve==2.5
sympy==1.13.1
textblob==0.18.0.post0
tifffile==2023.7.10
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
